+++ src/main/java/com/ferreteriafc/api/backend/persistence/repository/BrandRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ferreteriafc.api.backend.persistence.entity.Brand;
 
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Long> {
+public interface BrandRepository extends JpaRepository<Brand, Integer> {
 
     boolean existsByName(String name);
